fix proxy server settings and token missing issue
