+++ .env
@@ -1,20 +1,9 @@
-GOOGLE_CLIENT_ID=204710622072-mnff6851juhaqk2teuic35jsu81mgmv7.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=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
-
 MORALIS_API_KEY=IKkarA6u0V3JwTNgvjn1PwgzDbHZdg8JabdCqndcPhEO8ZcZcVTKEO3nZrDGra2L
-
 MORALIS_APP_DOMAIN=BeWater
-
-GITHUB_CLIENT_ID=6fd3d8d0476415d22382
-GITHUB_CLIENT_SECRET=ae19d64c109cadd8cf98ab9002a3d1dfb47d6eb4
-
-DISCORD_CLIENT_ID=1024206820660740106
-DISCORD_CLIENT_SECRET=RJGIRX_m62WRQDfH7-TcVdB3f3FQ4E_S
 
 NEXTAUTH_URL=http://localhost:3000/auth/connect-wallet
 NEXTAUTH_SECRET=68d9e5ea190d3da6e31417cd95a08450
 
 AUTH_REQUIRED=true
-MOCK_MODE=false
-NEXT_PUBLIC_BASE_PATH=/bw
-ENVIRONMENT=qa
+NEXT_PUBLIC_BASE_PATH=
+ENVIRONMENT=local
