+++ .env.proxy
@@ -1,0 +1,6 @@
+# Run bewater-website on this port, visit http://localhost:3001
+PROXY_SERVER_PORT=3001
+# Default proxy to Local Backend Server
+PROXY_TARGET='http://localhost:3000'
+# proxy to QA Backend Server (NOT working now!!!!!!!!)
+# PROXY_TARGET='http://bw-backend-elb-532860068.ap-southeast-1.elb.amazonaws.com',
