+++ next.config.js
@@ -3,7 +3,8 @@
 const api = {
   local: 'http://localhost:3000',
   qa: 'http://bw-backend-elb-532860068.ap-southeast-1.elb.amazonaws.com',
-  production: 'https://api.bewater.com',
+  // TODO: change it later
+  prod: 'https://api.bewater.xyz',
 };
 
 const basePath =
@@ -32,7 +33,6 @@ const nextConfig = {
     basePath,
     environment,
     authRequired: process.env.AUTH_REQUIRED === 'true',
-    mockMode: process.env.MOCK_MODE === 'true',
     apiHost: api[environment],
   },
   // Disabling Next.js ESLint check with custom one as there is
