+++ src/components/connect-wallet/connect-wallet.tsx
@@ -10,6 +10,7 @@ import {
   submitGetLoginMessage,
   submitVerifySignedMessage,
 } from '@/services/auth';
+import { urlWithBasePath } from '@/utils/urlWithBasePath';
 
 import type { UserLocalStorage } from '@/models/user';
 
@@ -34,9 +35,7 @@ export function ConnectWallet() {
             network: 'evm',
           };
           const { message } = await submitGetLoginMessage(messageParams);
-          // console.log({ message });
           const signature = await signMessageAsync({ message });
-          // console.log({ signature });
           const { token, userId, userProfile } =
             await submitVerifySignedMessage({
               message,
@@ -50,15 +49,15 @@ export function ConnectWallet() {
             isNewUser: !userProfile,
           });
           if (!userProfile) {
-            setTimeout(() => {
-              navigator.goToWelcome();
-            }, 1000);
+            window.location.href = urlWithBasePath('/welcome');
           } else {
-            navigator.goToUserSettings();
+            window.location.href = urlWithBasePath('/user/settings');
           }
         }
       } catch (error) {
         console.error(error);
+      } finally {
+        setIsLogining(false);
       }
     };
 
@@ -69,7 +68,6 @@ export function ConnectWallet() {
         navigator.goToUserSettings();
       }
     } else if (!isLogining && isConnected && !isAuthed(authToken)) {
-      // console.log('start asyncAuth...');
       setIsLogining(true);
       void asyncAuth();
     }
