+++ src/pages/welcome.page.tsx
@@ -7,7 +7,6 @@ import type { NextPage } from 'next';
 // Pass client to React Context Provider
 const PageWelcome: NextPage = () => {
   const token = useAuthContext();
-  console.log('welcome token: ', { token });
   return (
     <div className="flex flex-col h-[calc(100vh-160px)] justify-center items-center">
       <div className="w-[280px] flex flex-col">
