+++ src/services/helper/request.ts
@@ -1,19 +1,5 @@
-import getConfig from 'next/config';
-
+import { apiUrl } from '@/utils/apiUrl';
 import { withTimeout } from '@/utils/withTimeout';
-
-import type { NextRuntimeConfig } from '@/types/next-runtime-config';
-
-const {
-  publicRuntimeConfig: { apiHost, basePath },
-} = getConfig() as NextRuntimeConfig;
-
-export function apiUrl(path: string): string {
-  if (path.startsWith('/api') && apiHost.startsWith('http')) {
-    return `${apiHost}${path.replace('/api', '')}`;
-  }
-  return `${basePath}${path}`;
-}
 
 export async function fetchBody<T>(
   req: string,
